Refactor DISKS_TO_IGNORE usages
