--- src/middlewared/middlewared/plugins/disk_/enums.py
@@ -1,1 +1,0 @@
-DISKS_TO_IGNORE = ('sr', 'md', 'dm-', 'loop', 'zd')
