+++ src/middlewared/middlewared/plugins/disk_/retaste.py
@@ -7,6 +7,7 @@ import re
 from middlewared.schema import List, Str
 from middlewared.service import Service, accepts, job, private
 
+
 logger = logging.getLogger(__name__)
 
 SD_PATTERN = re.compile(r"^sd[a-z]+$")
